+++ packages/sphinx-test-reports/docs/conf.py
@@ -29,9 +29,9 @@ copyright = f"team useblocks, 2017-{now.year}"
 author = "team useblocks"
 
 # The short X.Y version
-version = "1.0"
+version = "1.1"
 # The full version, including alpha/beta/rc tags
-release = "1.0.2"
+release = "1.1.0"
 
 needs_id_regex = ".*"
 needs_css = "dark.css"