+++ packages/sphinx-test-reports/sphinxcontrib/test_reports/test_reports.py
@@ -30,7 +30,7 @@ else:
 
 # fmt: on
 
-VERSION = "1.0.2"
+VERSION = "1.1.0"
 
 
 def setup(app):